src/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
SQLAlchemy==1.4.31
PyMySQL==1.0.2
python-dotenv==0.21.0
prometheus_client==0.11.0